move some of RedirectionNotificationModal functions to its subcomponents
