+++ src/components/RedirectionNotificationForMobile.js
@@ -2,7 +2,26 @@ import * as s from "./RedirectionNotificationModal.sc";
 import CloseRoundedIcon from "@mui/icons-material/CloseRounded";
 import { Link } from "react-router-dom/cjs/react-router-dom";
 
-export default function RedirectionNotificationForMobile(props) {
+export default function RedirectionNotificationForMobile({
+  api,
+  article,
+  handleClose,
+  setIsArticleSaved,
+}) {
+  function handleSaveArticle() {
+    api.makePersonalCopy(article.id, (data) => {
+      if (data === "OK") {
+        setIsArticleSaved(true);
+      }
+    });
+  }
+
+  function handleSaveAndOpenArticle() {
+    handleSaveArticle();
+    // handleModalUse(); //Temporarily disabled for this function on mobile as it worked only when <Link> had its target set to _blank
+    handleClose();
+  }
+
   return (
     <>
       <s.Header>
@@ -16,7 +35,7 @@ export default function RedirectionNotificationForMobile(props) {
           <strong> Save and view the article</strong> to add it to your Saves.
         </p>
       </s.Body>
-      <s.CloseButton role="button" onClick={props.handleClose}>
+      <s.CloseButton role="button" onClick={handleClose}>
         <CloseRoundedIcon fontSize="medium" />
       </s.CloseButton>
       <s.Footer>
@@ -26,11 +45,8 @@ export default function RedirectionNotificationForMobile(props) {
               on mobile only when target="_blank". This issue didn't occur
               on the desktop and for regular <a> links. Needs further investigation
               if we want this functionality here  */}
-        <Link to={`/read/article?id=${props.article.id}`}>
-          <s.GoToArticleButton
-            role="button"
-            onClick={props.handleSaveAndOpenArticle}
-          >
+        <Link to={`/read/article?id=${article.id}`}>
+          <s.GoToArticleButton role="button" onClick={handleSaveAndOpenArticle}>
             Save and view the article
           </s.GoToArticleButton>
         </Link>
